bugfixin #1

+++ src/main/java/util/xml/XmlWrapper.java
@@ -10,12 +10,11 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
 
+import javax.xml.XMLConstants;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.FilenameUtils;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -24,10 +23,34 @@ import org.xml.sax.SAXException;
 
 import app.input.IOUtils;
 import app.input.InputFileException;
-import util.Logger;
 
 public class XmlWrapper
 {
+	public static DocumentBuilder newSecureDocumentBuilder() throws ParserConfigurationException
+	{
+		DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+		factory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
+		factory.setFeature("http://apache.org/xml/features/disallow-doctype-decl", true);
+		factory.setFeature("http://xml.org/sax/features/external-general-entities", false);
+		factory.setFeature("http://xml.org/sax/features/external-parameter-entities", false);
+		factory.setFeature("http://apache.org/xml/features/nonvalidating/load-external-dtd", false);
+
+		try {
+			factory.setAttribute(XMLConstants.ACCESS_EXTERNAL_DTD, "");
+			factory.setAttribute(XMLConstants.ACCESS_EXTERNAL_SCHEMA, "");
+			factory.setXIncludeAware(false);
+			factory.setExpandEntityReferences(false);
+		}
+		catch (IllegalArgumentException | UnsupportedOperationException e) {
+			ParserConfigurationException wrapped = new ParserConfigurationException(
+				"XML parser does not support the required security settings: " + e.getMessage());
+			wrapped.initCause(e);
+			throw wrapped;
+		}
+
+		return factory.newDocumentBuilder();
+	}
+
 	public static interface XmlSerializable
 	{
 		public void fromXML(XmlReader xmr, Element elem);
@@ -46,8 +69,7 @@ public class XmlWrapper
 			Document document;
 
 			try {
-				DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-				DocumentBuilder builder = factory.newDocumentBuilder();
+				DocumentBuilder builder = newSecureDocumentBuilder();
 				document = builder.parse(xmlFile);
 			}
 			catch (ParserConfigurationException e) {
@@ -70,8 +92,7 @@ public class XmlWrapper
 			Document document;
 
 			try {
-				DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-				DocumentBuilder builder = factory.newDocumentBuilder();
+				DocumentBuilder builder = newSecureDocumentBuilder();
 				document = builder.parse(stream);
 			}
 			catch (ParserConfigurationException e) {
@@ -541,7 +562,11 @@ public class XmlWrapper
 		public XmlWriter(File xmlFile, String ... headerComments) throws FileNotFoundException
 		{
 			try {
-				temp = File.createTempFile(FilenameUtils.getBaseName("StarRod_" + xmlFile.getName()), null);
+				File parent = xmlFile.getAbsoluteFile().getParentFile();
+				if (parent == null || !parent.isDirectory())
+					throw new IOException("Output directory does not exist for " + xmlFile);
+
+				temp = File.createTempFile("StarRod_", ".tmp", parent);
 				temp.deleteOnExit();
 
 				file = xmlFile;
@@ -559,39 +584,42 @@ public class XmlWrapper
 			}
 		}
 
-		public void save()
-		{
-			if (pw != null) {
-				pw.close();
-
-				try {
-					FileUtils.copyFile(temp, file);
-				}
-				catch (IOException e) {
-					Logger.printStackTrace(e);
-				}
-			}
-		}
-
-		/** Flushes and copies this writer's UTF-8 output or reports failure. */
-		public void saveOrThrow() throws IOException
+		public void save() throws IOException
 		{
 			if (pw == null)
 				throw new IOException("XML writer is closed");
-			pw.flush();
-			if (pw.checkError())
+
+			PrintWriter writer = pw;
+			writer.flush();
+			boolean writeFailed = writer.checkError();
+			writer.close();
+			writeFailed |= writer.checkError();
+			pw = null;
+
+			if (writeFailed)
 				throw new IOException("Could not write temporary XML output for " + file);
-			pw.close();
-			FileUtils.copyFile(temp, file);
+
+			IOUtils.atomicMoveFile(temp, file);
+			temp = null;
+		}
+
+		/** Compatibility alias for callers that explicitly requested checked save failures. */
+		public void saveOrThrow() throws IOException
+		{
+			save();
 		}
 
 		@Override
 		public void close()
 		{
-			if (temp != null)
-				temp.delete();
-			if (pw != null)
+			if (pw != null) {
 				pw.close();
+				pw = null;
+			}
+			if (temp != null) {
+				temp.delete();
+				temp = null;
+			}
 		}
 
 		public void printComment(String comment)
